+++ R/solve_seq.R
@@ -1,10 +1,9 @@
-#' @title Find subgame perfect equilibria of a sequential-form game
-#' @description \code{solve_seq()} finds subgame perfect equilibrium outcomes
-#'     of a sequential-form game.
-#'     This is a wrapper
-#'     function of \code{\link{solve_seq_matrix}}, \code{\link{solve_seq_char}}, and
+#' @title Find Nash equilibria of a sequential-form game
+#' @description \code{solve_seq()} finds Nash equilibrium outcomes of a
+#'     sequential-form game. This is a wrapper function of
+#'     \code{\link{solve_seq_matrix}}, \code{\link{solve_seq_char}}, and
 #'     \code{\link{solve_seq_fcn}}.
-#' @return A list containing subgame perfect equilibrium outcomes (if any), and
+#' @return A list containing Nash equilibrium outcomes (if any), and
 #'     the game table (if available).
 #' @inheritParams solve_seq_matrix
 #' @inheritParams solve_seq_char
@@ -13,7 +12,6 @@
 #' @export
 solve_seq <- function(game,
                       show_table = TRUE,
-                      cell_width = NULL,
                       mark_br = FALSE,
                       cons1 = NULL,
                       cons2 = NULL,
@@ -28,7 +26,6 @@ solve_seq <- function(game,
     solve_seq_matrix(game = game,
                      show_table = show_table,
                      mark_br = mark_br,
-                     cell_width = cell_width,
                      quietly = quietly)
 
   } else if (game$type == "char_function") {